add: valida_id
