+++ include/leitura.h
@@ -2,7 +2,7 @@
 //modulo leitura//
 //////////////////
 
-int ler_nome(char *variavel, int tam);
+void ler_nome(char *variavel, int tam);
 void ler_cpf(char *variavel, int tam);
 void ler_cell(char *variavel, int tam);
 void ler_n_quarto(char *variavel, int tam);
