+++ include/validacoes.h
@@ -7,3 +7,5 @@ int valida_num(char num);
 char* remove_char_especial(char* texto);
 int is_numeric(const char *num);
 int valida_cpf(char *CPF_original);
+int valida_id(char *ID);
+int valida_nome(char *nome);
