+++ object/leitura.c
@@ -5,44 +5,24 @@
 #include "utilidades.h"
 #include "validacoes.h"
 
-int valida_nome(char *nome){
-    int tamanho = strlen(nome);
-    int valido = (tamanho > 0);
-
-    if(!valido){
-        printf("NOME VAZIO!\n");
-        return False;
-    }
-
-    for(int i = 0; i < tamanho; i++){
-        if(!isalpha(nome[i]) && !isspace(nome[i])){
-            printf("NOME INVALIDO!\n");
-            return False;
-        }
-    }
-
-    return True;
-}
-
-int ler_nome(char *nome, int tam) {
+void ler_nome(char *variavel, int tam) {
     
     do{
-        input(nome, tam, "Digite o nome: ");
-    } while(!validar_nome(nome));
+        input(variavel, tam, "Digite o nome: ");
+    } while(!valida_nome(variavel));
 
-    printf("Nome valido!\n");
-    return True;
+    printf("\nNome valido!\n");
+
 }
-
-
-
 
 
 void ler_cpf(char *variavel, int tam){
+
     do {
         input(variavel, tam, "Digite o cpf: ");
     } while (!valida_cpf(variavel));
     
+    printf("\nCPF valido!\n");
 
 }
 
@@ -56,6 +36,10 @@ void ler_cell(char *variavel, int tam){
 
 void ler_n_quarto(char *variavel, int tam){
 
-    input(variavel, tam, "Digite o ID do quarto: ");
+    do {
+        input(variavel, tam, "Digite o ID do quarto: ");
+    } while (!valida_id(variavel));
     
+    printf("\nID valido!\n");
+
 }
